requirements.txt
pandas
numpy
streamlit
SQLAlchemy
mysql-connector-python
seaborn
matplotlib
plotly
Pillow
wordcloud
scipy
-e .